Reorganize metrics/trace exporter into files
